requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
oauth2client==4.1.3
Flask==2.3.3
gunicorn==21.2.0